Move ClusterSnapshot to upstreamsync
