+++ pkg/upstreamsync/snapshot/helpers.go
