+++ pkg/upstreamsync/snapshot/helpers_test.go
