+++ pkg/upstreamsync/snapshot/snapshot.go
