+++ pkg/upstreamsync/snapshot/snapshot_test.go
