+++ pkg/upstreamsync/snapshot/types.go
